requirements.txt
sqlalchemy>=2.0
psycopg[binary]>=3.1
pandas>=2.1
numpy>=1.26
scikit-learn>=1.4
joblib>=1.3
requests>=2.31
python-dotenv>=1.0
tqdm>=4.66
